Moved functions working on file system to dedicated module
